Add support for other kinds of drawer
