+++ README.md
@@ -53,7 +53,7 @@ Replace the main Layout of `DrawerLayout` with the `FullDraggableContainer` (or 
 ## TODO
 
 - [x] Add support for the right drawer / RTL
-- [ ] Add support for other kinds of drawer
+- [x] Add support for other kinds of drawer
 
 License
 -------
