+++ sample/src/main/java/com/drakeet/drawer/sample/MainActivity.kt
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.drakeet.drawer.sample
 
 import android.content.Context
